# system.management/src/main/java/com/bank/system/management/dto/CustomerResponse.java
package com.bank.system.management.dto;

import lombok.Builder;
import lombok.Data;

@Data
@Builder
public class CustomerResponse {
    private Long id;
    private String fullName;
    private String email;
    private String role;
}