# system.management/src/main/resources/openapi.yml
openapi: 3.1.0
info:
  title: Bank Management API
  description: JWT Authentication APIs
  version: '1.0'
servers:
  - url: 'http://localhost:8080'
    description: Generated server url
security:
  - Bearer Authentication: []
paths:
  '/api/accounts/{accountNumber}':
    get:
      tags:
        - Account Controller
      operationId: getAccount
      parameters:
        - name: accountNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AccountResponse'
    put:
      tags:
        - Account Controller
      operationId: updateProfile
      parameters:
        - name: accountNumber
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AccountResponse'
  /api/fd:
    post:
      tags:
        - Fixed Deposit APIs
      operationId: createFD
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFDRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/FDResponse'
  /api/auth/register:
    post:
      tags:
        - Authentication APIs
      summary: Register User
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/auth/login:
    post:
      tags:
        - Authentication APIs
      summary: Login User
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/accounts:
    post:
      tags:
        - Account Controller
      operationId: createAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AccountResponse'
  '/api/accounts/{accountNumber}/debit':
    post:
      tags:
        - Account Controller
      operationId: debitAmount
      parameters:
        - name: accountNumber
          in: path
          required: true
          schema:
            type: string
        - name: amount
          in: query
          required: true
          schema:
            type: number
            format: double
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AccountResponse'
  '/api/accounts/{accountNumber}/credit':
    post:
      tags:
        - Account Controller
      operationId: creditAmount
      parameters:
        - name: accountNumber
          in: path
          required: true
          schema:
            type: string
        - name: amount
          in: query
          required: true
          schema:
            type: number
            format: double
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AccountResponse'
  '/api/transactions/{accountNumber}':
    get:
      tags:
        - Transaction Dashboard APIs
      operationId: getTransactions
      parameters:
        - name: accountNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionResponse'
  /api/transactions/search:
    get:
      tags:
        - Transaction Dashboard APIs
      operationId: searchTransaction
      parameters:
        - name: transactionId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TransactionResponse'
  /api/transactions/date-range:
    get:
      tags:
        - Transaction Dashboard APIs
      operationId: getTransactionsByDateRange
      parameters:
        - name: accountNumber
          in: query
          required: true
          schema:
            type: string
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionResponse'
  '/api/transactions/dashboard/{accountNumber}':
    get:
      tags:
        - Transaction Dashboard APIs
      operationId: getDashboard
      parameters:
        - name: accountNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DashboardResponse'
  '/api/fd/{fdNumber}':
    get:
      tags:
        - Fixed Deposit APIs
      operationId: getFDDetails
      parameters:
        - name: fdNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/FDResponse'
  /api/fd/my-fds:
    get:
      tags:
        - Fixed Deposit APIs
      operationId: getMyFds
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FDResponse'
  /api/fd/calculator:
    get:
      tags:
        - Fixed Deposit APIs
      operationId: calculate
      parameters:
        - name: amount
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: tenure
          in: query
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/FDCalculatorResponse'
  /api/customer/profile:
    get:
      tags:
        - Customer Controller
      operationId: profile
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomerResponse'
  /api/admin/customers:
    get:
      tags:
        - Admin Controller
      operationId: getAllCustomers
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerResponse'
  '/api/accounts/{accountNumber}/balance':
    get:
      tags:
        - Account Controller
      operationId: getBalance
      parameters:
        - name: accountNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: number
components:
  schemas:
    UpdateProfileRequest:
      type: object
      properties:
        mobileNumber:
          type: string
        address:
          type: string
    AccountResponse:
      type: object
      properties:
        accountNumber:
          type: string
        customerName:
          type: string
        email:
          type: string
        accountType:
          type: string
          enum:
            - SAVINGS
            - CURRENT
        branchName:
          type: string
        openingDate:
          type: string
          format: date
        currentBalance:
          type: number
        accountStatus:
          type: string
          enum:
            - ACTIVE
            - INACTIVE
            - CLOSED
        mobileNumber:
          type: string
        address:
          type: string
    CreateFDRequest:
      type: object
      properties:
        depositAmount:
          type: number
          format: double
          minimum: 10000.01
        tenureYears:
          type: integer
          format: int32
      required:
        - depositAmount
        - tenureYears
    FDResponse:
      type: object
      properties:
        fdNumber:
          type: string
        depositAmount:
          type: number
          format: double
        interestRate:
          type: number
          format: double
        tenureYears:
          type: integer
          format: int32
        depositDate:
          type: string
          format: date
        maturityDate:
          type: string
          format: date
        maturityAmount:
          type: number
          format: double
    RegisterRequest:
      type: object
      properties:
        fullName:
          type: string
          maxLength: 50
          minLength: 3
        email:
          type: string
          format: email
          minLength: 1
          pattern: '^[A-Za-z0-9._%+-]+@gmail\.com$'
        password:
          type: string
          maxLength: 20
          minLength: 8
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&]).+$'
        role:
          type: string
          enum:
            - ADMIN
            - CUSTOMER
      required:
        - email
        - fullName
        - password
        - role
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          minLength: 1
        password:
          type: string
          minLength: 1
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
    AccountRequest:
      type: object
      properties:
        customerName:
          type: string
        email:
          type: string
        accountType:
          type: string
          enum:
            - SAVINGS
            - CURRENT
        branchName:
          type: string
        mobileNumber:
          type: string
        address:
          type: string
    TransactionResponse:
      type: object
      properties:
        transactionId:
          type: string
        accountNumber:
          type: string
        transactionDate:
          type: string
          format: date-time
        transactionType:
          type: string
        transactionAmount:
          type: number
          format: double
        availableBalance:
          type: number
          format: double
    DashboardResponse:
      type: object
      properties:
        totalCredit:
          type: number
          format: double
        totalDebit:
          type: number
          format: double
        currentBalance:
          type: number
        lastTenTransactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionResponse'
    FDCalculatorResponse:
      type: object
      properties:
        depositAmount:
          type: number
          format: double
        interestRate:
          type: number
          format: double
        tenureYears:
          type: integer
          format: int32
        maturityAmount:
          type: number
          format: double
    CustomerResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
  securitySchemes:
    Bearer Authentication:
      type: http
      scheme: bearer
      bearerFormat: JWT